requirements.txt
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
ttkthemes==3.2.2
